[TIR Pass] decouple flatten buffer to lower opaque block pass and flatten buffer.
